backend/pyproject.toml
[project]
name = "blank"
version = "0.1.0"
authors = [
    { name = "Faiz Surani", email = "faiz@faizsurani.com" },
]
description = "A blank project"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
]
requires-python = "~=3.11.10"
dependencies = [
    "fastapi[standard]>=0.114.2",
    "pydantic>=2.9.1",
    "rl",
    "psycopg2-binary>=2.9.10",
    "sqlalchemy>=2.0.36",
    "alembic>=1.14.0",
    "tenacity>=9.0.0",
    "httpx>=0.27.2",
    "gunicorn>=23.0.0",
    "boto3>=1.38.41",
    "celery>=5.5.3",
    "redis>=6.2.0",
    "alembic-utils>=0.8.8",
    "structlog>=25.4.0",
    "sentry-sdk[celery,fastapi,sqlalchemy]>=2.30.0",
]

[dependency-groups]
dev = [
    "factory-boy>=3.3.3",
    "ipython>=8.27.0",
    "pgcli>=4.3.0",
    "pre-commit>=3.8.0",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
    "pytest-postgresql>=7.0.2",
    "ruff>=0.12.0",
]

[tool.uv.sources]
rl = { git = "https://github.com/ProbablyFaiz/rl.git" }

[project.urls]
homepage = "https://github.com/ProbablyFaiz/blank"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 88

[tool.ruff.lint]
extend-select = ["B", "E", "I", "N", "PTH", "COM", "C4", "UP"]
extend-ignore = ["COM812", "E712", "E501", "B904"]

[tool.pytest.ini_options]
testpaths = ["test"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
]
markers = [
    "slow: marks tests as slow",
    "integration: marks tests as integration tests",
]
